+++ packages/atomic/vitest-utils/testing-helpers/fixtures/headless/search/result-list-controller.ts
@@ -1,4 +1,5 @@
 import type {ResultList, ResultListState} from '@coveo/headless';
+import {vi} from 'vitest';
 import {genericSubscribe} from '../common.js';
 
 export const defaultState = {
@@ -9,12 +10,13 @@ export const defaultState = {
   results: [],
   searchResponseId: 'test-search-response-id',
   moreResultsAvailable: false,
-};
+} satisfies ResultListState;
 
 export const defaultImplementation = {
   subscribe: genericSubscribe,
   state: defaultState,
-};
+  fetchMoreResults: vi.fn(),
+} satisfies ResultList;
 
 export const buildFakeResultList = ({
   implementation,